+++ gin/sentrygin.go
@@ -9,20 +9,30 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-type Handler struct {
+const valuesKey = "sentry"
+
+type handler struct {
 	repanic         bool
 	waitForDelivery bool
 	timeout         time.Duration
 }
 
 type Options struct {
-	Repanic         bool
+	// Repanic configures whether Sentry should repanic after recovery, in most cases it should be set to true,
+	// as gin.Default includes it's own Recovery middleware what handles http responses.
+	Repanic bool
+	// WaitForDelivery configures whether you want to block the request before moving forward with the response.
+	// Because Gin's default `Recovery` handler doesn't restart the application,
+	// it's safe to either skip this option or set it to `false`.
 	WaitForDelivery bool
-	Timeout         time.Duration
+	// Timeout for the event delivery requests.
+	Timeout time.Duration
 }
 
+// New returns a function that satisfies gin.HandlerFunc interface
+// It can be used with Use() methods.
 func New(options Options) gin.HandlerFunc {
-	handler := Handler{
+	handler := handler{
 		repanic:         false,
 		timeout:         time.Second * 2,
 		waitForDelivery: false,
@@ -36,29 +46,23 @@ func New(options Options) gin.HandlerFunc {
 		handler.waitForDelivery = true
 	}
 
-	return handler.handle()
+	return handler.handle
 }
 
-func (h *Handler) handle() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		r := c.Copy().Request
-		ctx := sentry.SetHubOnContext(
-			context.WithValue(r.Context(), sentry.RequestContextKey, r),
-			sentry.CurrentHub().Clone(),
-		)
-		defer h.recoverWithSentry(ctx, r)
-		c.Request = r.WithContext(ctx)
-		c.Next()
-	}
+func (h *handler) handle(ctx *gin.Context) {
+	hub := sentry.CurrentHub().Clone()
+	ctx.Set(valuesKey, hub)
+	defer h.recoverWithSentry(hub, ctx.Request)
+	ctx.Next()
 }
 
-func (h *Handler) recoverWithSentry(ctx context.Context, r *http.Request) {
+func (h *handler) recoverWithSentry(hub *sentry.Hub, r *http.Request) {
 	if err := recover(); err != nil {
-		hub := sentry.GetHubFromContext(ctx)
-		hub.ConfigureScope(func(scope *sentry.Scope) {
-			scope.SetRequest(sentry.Request{}.FromHTTPRequest(r))
-		})
-		eventID := hub.RecoverWithContext(ctx, err)
+		hub.Scope().SetRequest(sentry.Request{}.FromHTTPRequest(r))
+		eventID := hub.RecoverWithContext(
+			context.WithValue(r.Context(), sentry.RequestContextKey, r),
+			err,
+		)
 		if eventID != nil && h.waitForDelivery {
 			hub.Flush(h.timeout)
 		}
@@ -66,4 +70,14 @@ func (h *Handler) recoverWithSentry(ctx context.Context, r *http.Request) {
 			panic(err)
 		}
 	}
+}
+
+// GetHubFromContext retrieves attached *sentry.Hub instance from iris.Context
+func GetHubFromContext(ctx *gin.Context) *sentry.Hub {
+	if hub, ok := ctx.Get(valuesKey); ok {
+		if hub, ok := hub.(*sentry.Hub); ok {
+			return hub
+		}
+	}
+	return nil
 }