+++ echo/sentryecho.go
@@ -1,4 +1,4 @@
-package sentrygin
+package sentryecho
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/getsentry/sentry-go"
-	"github.com/gin-gonic/gin"
+	"github.com/labstack/echo"
 )
 
 const valuesKey = "sentry"
@@ -19,19 +19,19 @@ type handler struct {
 
 type Options struct {
 	// Repanic configures whether Sentry should repanic after recovery, in most cases it should be set to true,
-	// as gin.Default includes it's own Recovery middleware what handles http responses.
+	// as echo includes it's own Recover middleware what handles http responses.
 	Repanic bool
 	// WaitForDelivery configures whether you want to block the request before moving forward with the response.
-	// Because Gin's default `Recovery` handler doesn't restart the application,
+	// Because Echo's `Recover` handler doesn't restart the application,
 	// it's safe to either skip this option or set it to `false`.
 	WaitForDelivery bool
 	// Timeout for the event delivery requests.
 	Timeout time.Duration
 }
 
-// New returns a function that satisfies gin.HandlerFunc interface
+// New returns a function that satisfies echo.HandlerFunc interface
 // It can be used with Use() methods.
-func New(options Options) gin.HandlerFunc {
+func New(options Options) echo.MiddlewareFunc {
 	handler := handler{
 		repanic:         false,
 		timeout:         time.Second * 2,
@@ -49,11 +49,13 @@ func New(options Options) gin.HandlerFunc {
 	return handler.handle
 }
 
-func (h *handler) handle(ctx *gin.Context) {
-	hub := sentry.CurrentHub().Clone()
-	ctx.Set(valuesKey, hub)
-	defer h.recoverWithSentry(hub, ctx.Request)
-	ctx.Next()
+func (h *handler) handle(next echo.HandlerFunc) echo.HandlerFunc {
+	return func(ctx echo.Context) error {
+		hub := sentry.CurrentHub().Clone()
+		ctx.Set(valuesKey, hub)
+		defer h.recoverWithSentry(hub, ctx.Request())
+		return next(ctx)
+	}
 }
 
 func (h *handler) recoverWithSentry(hub *sentry.Hub, r *http.Request) {
@@ -72,12 +74,10 @@ func (h *handler) recoverWithSentry(hub *sentry.Hub, r *http.Request) {
 	}
 }
 
-// GetHubFromContext retrieves attached *sentry.Hub instance from gin.Context.
-func GetHubFromContext(ctx *gin.Context) *sentry.Hub {
-	if hub, ok := ctx.Get(valuesKey); ok {
-		if hub, ok := hub.(*sentry.Hub); ok {
-			return hub
-		}
+// GetHubFromContext retrieves attached *sentry.Hub instance from echo.Context.
+func GetHubFromContext(ctx echo.Context) *sentry.Hub {
+	if hub, ok := ctx.Get(valuesKey).(*sentry.Hub); ok {
+		return hub
 	}
 	return nil
 }